fix: wrong type for id

+++ pkg/domain/domain.go
@@ -1,7 +1,7 @@
 package domain
 
 type User struct {
-	Id             int64
+	Id             int32
 	Username       string
 	Password       string
 	ExpirationDays int32 //in days
